Refactor Delegated Functions into mini-loop
It was developed in isolation and subsequent tickets
identified a way of tidying the provider controllers
to make the root folder slightly less congested
