+++ app/helpers/proceeding_helper.rb
@@ -4,9 +4,9 @@ module ProceedingHelper
     proceedings = application.proceedings.in_order_of_addition
     if proceedings.count > 1
       position = proceedings.pluck(:id).find_index(proceeding.id) + 1
-      I18n.t("providers.delegated_functions.show.multi_proceeding", position:, total: proceedings.count)
+      I18n.t("providers.proceeding_loop.multi_proceeding", position:, total: proceedings.count)
     elsif proceedings.count == 1
-      I18n.t("providers.delegated_functions.show.single_proceeding")
+      I18n.t("providers.proceeding_loop.single_proceeding")
     end
   end
 end
