+++ app/views/providers/proceeding_loop/delegated_functions/show.html.erb
