+++ config/routes.rb
@@ -224,7 +224,7 @@ Rails.application.routes.draw do
       resource :no_eligibility_assessment, only: %i[show update]
       resource :used_multiple_delegated_functions, only: %i[show update] # TODO: Delete after mini loop implemented
       resource :confirm_multiple_delegated_functions, only: %i[show update] # TODO: Refactor after mini loop
-      resources :delegated_functions, only: %i[show update]
+      resources :delegated_functions, only: %i[show update], controller: "proceeding_loop/delegated_functions"
       resources :client_involvement_type, only: %i[show update], controller: "proceeding_loop/client_involvement_type"
       resource :use_ccms, only: %i[show]
       resources :use_ccms_employed, only: %i[index]
